refactor: useApplicationAutoSave 훅으로 임시 저장 로직 분리
